Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

CaNNOLeS.CaNNOLeSSolver

CaNNOLeS.cannoles

CaNNOLeS.newton_system!

CaNNOLeS.try_to_factorize

​

CaNNOLeS.CaNNOLeSSolver — Type

cannoles(nls)

Implementation of a solver for Nonlinear Least Squares with nonlinear constraints.

$min f(x) = ¹/₂‖F(x)‖² s.t. c(x) = 0$

For advanced usage, first define a CaNNOLeSSolver to preallocate the memory used in the algorithm, and then call solve!:

solver = CaNNOLeSSolver(nls)
solve!(solver, nls; kwargs...)

or even pre-allocate the output:

stats = GenericExecutionStats(nls)
solve!(solver, nls, stats; kwargs...)

Arguments

nls :: AbstractNLSModel: nonlinear least-squares model created using NLPModels.

Keyword arguments

x::AbstractVector = nls.meta.x0: the initial guess;

λ::AbstractVector = eltype(x)[]: the initial Lagrange multiplier;

method::Symbol = :Newton: available methods :Newton, :LM, :Newton_noFHess, and :Newton_vanishing;

linsolve::Symbol = :ma57: solver to compute LDLt factorization. Available methods are: :ma57, :ldlfactorizations;

max_f::Real = 100000: maximum number of evaluations computed by sum_counters(nls);

max_time::Float64 = 30.0: maximum time limit in seconds;

max_inner::Int = 10000: maximum number of inner iterations;

ϵtol::Real = √eps(eltype(x)): stopping tolerance;

verbose::Int = 0: if > 0, display iteration details every verbose iteration;

check_small_residual::Bool = true: if true, stop whenever $‖F(x)‖₂² ≤ ϵtol$ and $‖c(xᵏ)‖∞ ≤ √ϵtol$;

always_accept_extrapolation::Bool = false: if true, run even if the extrapolation step fails;

δdec::Real = eltype(x)(0.1): reducing factor on the parameter δ.

The algorithm stops when $‖c(xᵏ)‖∞ ≤ ϵtol$ and $‖∇F(xᵏ)ᵀF(xᵏ) - ∇c(xᵏ)ᵀλᵏ‖ ≤ ϵtol * max(1, ‖λᵏ‖ / 100ncon)$.

Output

The value returned is a GenericExecutionStats, see SolverCore.jl.

Examples

using CaNNOLeS, ADNLPModels
nls = ADNLSModel(x -> x, ones(3), 3)
stats = cannoles(nls, linsolve = :ldlfactorizations, verbose = 0)
stats

# output

"Execution stats: first-order stationary"

using CaNNOLeS, ADNLPModels
nls = ADNLSModel(x -> x, ones(3), 3)
solver = CaNNOLeSSolver(nls)
stats = solve!(solver, nls, linsolve = :ldlfactorizations, verbose = 0)
stats

# output

"Execution stats: first-order stationary"

CaNNOLeS.cannoles — Method

cannoles(nls)

Implementation of a solver for Nonlinear Least Squares with nonlinear constraints.

$min f(x) = ¹/₂‖F(x)‖² s.t. c(x) = 0$

For advanced usage, first define a CaNNOLeSSolver to preallocate the memory used in the algorithm, and then call solve!:

solver = CaNNOLeSSolver(nls)
solve!(solver, nls; kwargs...)

or even pre-allocate the output:

stats = GenericExecutionStats(nls)
solve!(solver, nls, stats; kwargs...)

Arguments

nls :: AbstractNLSModel: nonlinear least-squares model created using NLPModels.

Keyword arguments

x::AbstractVector = nls.meta.x0: the initial guess;

λ::AbstractVector = eltype(x)[]: the initial Lagrange multiplier;

method::Symbol = :Newton: available methods :Newton, :LM, :Newton_noFHess, and :Newton_vanishing;

linsolve::Symbol = :ma57: solver to compute LDLt factorization. Available methods are: :ma57, :ldlfactorizations;

max_f::Real = 100000: maximum number of evaluations computed by sum_counters(nls);

max_time::Float64 = 30.0: maximum time limit in seconds;

max_inner::Int = 10000: maximum number of inner iterations;

ϵtol::Real = √eps(eltype(x)): stopping tolerance;

verbose::Int = 0: if > 0, display iteration details every verbose iteration;

check_small_residual::Bool = true: if true, stop whenever $‖F(x)‖₂² ≤ ϵtol$ and $‖c(xᵏ)‖∞ ≤ √ϵtol$;

always_accept_extrapolation::Bool = false: if true, run even if the extrapolation step fails;

δdec::Real = eltype(x)(0.1): reducing factor on the parameter δ.

The algorithm stops when $‖c(xᵏ)‖∞ ≤ ϵtol$ and $‖∇F(xᵏ)ᵀF(xᵏ) - ∇c(xᵏ)ᵀλᵏ‖ ≤ ϵtol * max(1, ‖λᵏ‖ / 100ncon)$.

Output

The value returned is a GenericExecutionStats, see SolverCore.jl.

Examples

using CaNNOLeS, ADNLPModels
nls = ADNLSModel(x -> x, ones(3), 3)
stats = cannoles(nls, linsolve = :ldlfactorizations, verbose = 0)
stats

# output

"Execution stats: first-order stationary"

using CaNNOLeS, ADNLPModels
nls = ADNLSModel(x -> x, ones(3), 3)
solver = CaNNOLeSSolver(nls)
stats = solve!(solver, nls, linsolve = :ldlfactorizations, verbose = 0)
stats

# output

"Execution stats: first-order stationary"

CaNNOLeS.newton_system! — Method

newton_system!(d, nvar, nequ, ncon, rhs, LDLT, ρold, params)

Compute an LDLt factorization of the (nvar + nequ + ncon)-square matrix for the Newton system contained in LDLT, i.e., sparse(LDLT.rows, LDLT.cols, LDLT.vals, N, N). If the factorization fails, a new factorization is attempted with an increased value for the regularization ρ as long as it is smaller than params[:ρmax]. The factorization is then used to solve the linear system whose right-hand side is rhs.

Output

d: the solution of the linear system;

solve_success: true if the usage of the LDLt factorization is successful;

ρ: the value of the regularization parameter used in the factorization;

ρold: the value of the regularization parameter used in the previous successful factorization, or 0 if this is the first one;

nfact: the number of factorization attempts.

CaNNOLeS.try_to_factorize — Function

success = try_to_factorize(LDLT::LinearSolverStruct, nvar::Integer, nequ::Integer, ncon::Integer, eig_tol::Real)

Compute the LDLt factorization of A = sparse(LDLT.rows, LDLT.cols, LDLT.vals, N, N) where N = nvar + ncon + nequ and return true in case of success.